Renamed package to "xml"
