+++ src/main/java/com/datasonnet/xml/BadgerFishXMLStreamWriter.java
@@ -1,4 +1,4 @@
-package com.datasonnet.badgerfish;
+package com.datasonnet.xml;
 
 import org.codehaus.jettison.AbstractXMLStreamWriter;
 import org.codehaus.jettison.Node;
